refactor: chat FSD 폴더 구조 맞춰서 폴더 및 파일명 변경
