+++ src/features/send-chat/ui/index.ts
@@ -1,0 +1,1 @@
+export { SendChatForm } from './send-chat-from';
