+++ src/features/send-chat/ui/send-chat-from/index.ts
@@ -1,0 +1,1 @@
+export { SendChatForm } from './send-chat-form';
