+++ src/pages/chat/ui/chat-page/index.ts
@@ -1,0 +1,1 @@
+export { ChatPage } from './chat-page';
